added js docs and changed modules from .js to .,mjs
