+++ index.html
@@ -39,6 +39,7 @@
 
         </footer>
         <script src="./node_modules/bootstrap//dist/js/bootstrap.bundle.js"></script>
+        <script src="./js/index.js" type="module"></script>
         
     </body>
