--- js/api/storage/clear.js
@@ -1,3 +1,0 @@
-export function clear(key) {
-    localStorage.clear();
-}
