+++ js/api/storage/clear.mjs
@@ -1,0 +1,11 @@
+/**
+ * this clears the whole local storage
+ * @param {*} key 
+ * 
+ * ```
+ * clear(key);
+ * ```
+ */
+export function clear(key) {
+    localStorage.clear();
+}
