--- js/api/storage/index.js
@@ -1,11 +1,0 @@
-import { save } from "./save";
-import { remove } from "./remove";
-import { clear } from "./clear";
-import { load } from "./load";
-
-export default {
-    save,
-    remove,
-    clear,
-    load,
-}
