+++ js/api/storage/index.mjs
@@ -1,0 +1,11 @@
+import { save } from "./save.mjs";
+import { remove } from "./remove.mjs";
+import { clear } from "./clear.mjs";
+import { load } from "./load.mjs";
+
+export default {
+    save,
+    remove,
+    clear,
+    load,
+}
