+++ js/api/storage/load.mjs
@@ -1,3 +1,12 @@
+/**
+ * returns information from local storage based on the key
+ * @param {*} key 
+ * @returns 
+ * ```
+ * load(key);
+ * ```
+ */
+
 export function load(key) {
     try {
         const serializedState = localStorage.getItem(key);
