--- js/api/storage/remove.js
@@ -1,3 +1,0 @@
-export function remove(key) {
-    localStorage.removeItem(key);
-}
