+++ js/api/storage/remove.mjs
@@ -1,0 +1,10 @@
+/**
+ * this will remove one item from local storage based on key
+ * @param {*} key
+ * ```
+ * remove(key);
+ * ```
+ */
+export function remove(key) {
+    localStorage.removeItem(key);
+}
