--- js/api/storage/save.js
@@ -1,4 +1,0 @@
-export function save (key, state) {
-    const serializedState = JSON.stringify(state);
-    localStorage.setItem(key, serializedState);
-}
