+++ js/api/variables/index.mjs
@@ -1,0 +1,5 @@
+import { url } from "./url.mjs";
+
+export default {
+    url,
+}
