+++ js/api/variables/url.mjs
@@ -1,0 +1,1 @@
+export const url = "https://api.noroff.dev/api/v1/social/";
